# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IPO Alert API",
    "version": "2.0.0",
    "description": "A comprehensive API for Nepal stock market data, IPO tracking, portfolio management, and price alerts",
    "contact": {
      "name": "IPO Alert Team",
      "email": "support@ipo-alert.com"
    }
  },
  "servers": [
    {
      "url": "https://ipo-alert-latest.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "email": {"type": "string", "format": "email"},
          "name": {"type": "string"},
          "profilePicture": {"type": "string"},
          "authProvider": {"type": "string", "enum": ["local", "google"]},
          "lastLogin": {"type": "string", "format": "date-time"},
          "createdAt": {"type": "string", "format": "date-time"}
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "token": {"type": "string"},
          "user": {"$ref": "#/components/schemas/User"},
          "error": {"type": "string"}
        }
      },
      "PortfolioHolding": {
        "type": "object",
        "properties": {
          "_id": {"type": "string"},
          "userId": {"type": "string"},
          "ticker": {"type": "string"},
          "quantity": {"type": "number"},
          "averagePrice": {"type": "number"},
          "purchaseDate": {"type": "string", "format": "date"},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
      "PriceAlert": {
        "type": "object",
        "properties": {
          "_id": {"type": "string"},
          "userId": {"type": "string"},
          "ticker": {"type": "string"},
          "targetPrice": {"type": "number"},
          "condition": {"type": "string", "enum": ["above", "below"]},
          "type": {"type": "string", "enum": ["one-time", "recurring"]},
          "triggered": {"type": "boolean"},
          "lastTriggered": {"type": "string", "format": "date-time"},
          "createdAt": {"type": "string", "format": "date-time"}
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": false},
          "error": {"type": "string"}
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string", "minLength": 6},
                  "name": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AuthResponse"}
              }
            }
          },
          "400": {
            "description": "Registration failed",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AuthResponse"}
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
    "/api/auth/google": {
      "post": {
        "summary": "Google OAuth login",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["idToken"],
                "properties": {
                  "idToken": {"type": "string", "description": "Google ID token"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google login successful",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AuthResponse"}
              }
            }
          },
          "400": {
            "description": "Invalid Google token",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "summary": "Get user profile",
        "tags": ["Authentication"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "user": {"$ref": "#/components/schemas/User"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
    "/api/portfolio": {
      "get": {
        "summary": "Get user's portfolio holdings",
        "tags": ["Portfolio"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Portfolio holdings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/PortfolioHolding"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add portfolio holding",
        "tags": ["Portfolio"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ticker", "quantity", "averagePrice", "purchaseDate"],
                "properties": {
                  "ticker": {"type": "string"},
                  "quantity": {"type": "number"},
                  "averagePrice": {"type": "number"},
                  "purchaseDate": {"type": "string", "format": "date"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Portfolio holding added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {"$ref": "#/components/schemas/PortfolioHolding"},
                    "message": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/{id}": {
      "put": {
        "summary": "Update portfolio holding",
        "tags": ["Portfolio"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {"type": "number"},
                  "averagePrice": {"type": "number"},
                  "purchaseDate": {"type": "string", "format": "date"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Portfolio holding updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {"$ref": "#/components/schemas/PortfolioHolding"},
                    "message": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete portfolio holding",
        "tags": ["Portfolio"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio holding deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "message": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/price-alerts": {
      "get": {
        "summary": "Get user's price alerts",
        "tags": ["Price Alerts"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Price alerts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/PriceAlert"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create price alert",
        "tags": ["Price Alerts"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ticker", "targetPrice", "condition"],
                "properties": {
                  "ticker": {"type": "string"},
                  "targetPrice": {"type": "number"},
                  "condition": {"type": "string", "enum": ["above", "below"]},
                  "type": {"type": "string", "enum": ["one-time", "recurring"], "default": "one-time"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price alert created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {"$ref": "#/components/schemas/PriceAlert"},
                    "message": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}